Move prometheus reconciler logic to its own
